[Move]Command move to client.rs
