+++ src/discord.rs
@@ -1,3 +1,2 @@
 pub mod action;
 pub mod client;
-pub mod command;
